% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement.R
\name{agreement}
\alias{agreement}
\title{Measuring Agreement between a New Device and Multiple Imperfect Reference Standards}
\usage{
agreement(
  references = "rater1 + rater2 + rater3",
  device = "CAD",
  index = c("CCC", "TDI"),
  transformation = function(x) 1/x,
  data = data,
  p = 0.05,
  boot = NULL
)
}
\arguments{
\item{device}{the column name for device.}

\item{index}{the agreement index. Can be "CCC" or "TDI"}

\item{data}{a dataframe containing raters and CAD}

\item{p}{available when using TDI; indicating the fraction of items with differences numerically exceeding the TDI}

\item{reference}{a references 'formula'. The expression 'rater1 + rater2 + rater3' indicates that three raters are being used, with the column names being 'rater1', 'rater2', and 'rater3'.}

\item{tranformation}{the transformation function g(x)}
}
\value{
a list of estimates, standard error, and confidence interval
}
\description{
Measuring Agreement between a New Device and Multiple Imperfect Reference Standards
}
